removeBG/requirements.txt
flask>=2.0.0
rembg>=2.0.0
pillow>=8.0.0
onnxruntime>=1.10.0